+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package are documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2021-??-??
+### Changed
+ - Improved message when the current selection doesn't contain a Component or ScriptableObject with a [PlayModeInspectorMethod] attribute.
+
 ## [1.2.0] - 2020-08-13
 ### Changed
  - Removed "sealed" keyword from the PlayModeInspectorMethodAttribute class. This allows to derive from it and use your own attribute in your code. In case you want to get rid of the PlayMode Inspector package, you only need to change your own attribute and everything still compiles.